Update samples/js so it works again
samples/js was broken after I switched it to Node 7.x because EventEmitter.prototype was removed.

Due to some dependency weirdness, I actually had to add an npm-shrinkwrap.json file to ensure that we would load a new enough version of socket.io that was compatible with 7.x.

While I was in here, I also removed the junit reporter (no idea why I had it here in the first place).

Fixes #23